move nat tests from container's unit test to nat's ones
